refactor: cleanup move to poetry
