+++ pyproject.toml
@@ -13,6 +13,9 @@ pytest = "5.4.3"
 pytest-cov = "2.8.1"
 tox = "3.14.4"
 
+[tool.poetry.scripts]
+testsuite = 'testsuite.console:main'
+
 [build-system]
 requires = ["poetry>=0.12"]
 build-backend = "poetry.masonry.api"
