+++ tox.ini
@@ -19,7 +19,7 @@ extras =
 [testenv:generated_sch]
 whitelist_externals = poetry
 commands =
-    poetry run python ./buildingsch.py generate_all --dry-run
+    poetry run testsuite generate_all --dry-run
 
 [testenv:pre-commit]
 deps =
